refactor: rename to AnnualReport.js
